Fixed a bug in position offset correction

+++ src/Source.h
@@ -70,7 +70,7 @@ PUBLIC long int      Source_get_par_int         (const Source *self, const size_
 PUBLIC double        Source_get_par_by_name_flt (const Source *self, const char *name);
 PUBLIC long int      Source_get_par_by_name_int (const Source *self, const char *name);
 
-PUBLIC void          Source_offset_xyz          (Source *self, const long int dx, const long int dy, const long int dz);
+PUBLIC void          Source_offset_xyz          (Source *self, const size_t dx, const size_t dy, const size_t dz);
 
 PUBLIC bool          Source_par_exists          (const Source *self, const char *name, size_t *index);
